[WIP] dcore + checkin
