+++ Cargo.toml
@@ -22,6 +22,7 @@ repository = "https://github.com/denoland/deno_core"
 deno_core = { version = "0.278.0", path = "./core" }
 deno_ops = { version = "0.154.0", path = "./ops" }
 serde_v8 = { version = "0.187.0", path = "./serde_v8" }
+deno_core_testing = { path = "./testing" }
 
 v8 = { version = "0.91.0", default-features = false }
 deno_ast = { version = "=0.35.3", features = ["transpiling"] }
