+++ dcore/Cargo.toml
@@ -14,8 +14,12 @@ description = "A simple binary that builds on deno_core"
 [[bin]]
 name = "dcore"
 
+[build-dependencies]
+deno_core_testing.workspace = true
+
 [dependencies]
 anyhow.workspace = true
+deno_core_testing.workspace = true
 clap = "4.5.1"
 deno_core.workspace = true
 fastwebsockets = { version = "0.6", features = ["upgrade", "unstable-split"] }
